intermediate commit for backend

+++ backend/src/main/java/com/company/holikov/backend/config/jwt/JwtTokenProvider.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 @Component
@@ -17,7 +18,7 @@ public class JwtTokenProvider {
     private String secretKey;
 
     @Value("${app.jwtExpirationMs}")
-    private int jwtExpirationMs;
+    private long jwtExpirationMs;
 
     public String generateJwtToken(Authentication authentication) {
